apply max blobs per block during validation
deneb and electra will still be using their constants, fulu will be using the schedule as this stands.

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/BlobSchedule.java
@@ -15,4 +15,4 @@ package tech.pegasys.teku.spec.config;
 
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public record BlobSchedule(UInt64 epoch, UInt64 maxBlobsPerBlock) {}
+public record BlobSchedule(UInt64 epoch, int maxBlobsPerBlock) {}
